feat: перенести инфраструктуру по работе с запросами предметов в модуль server
