+++ shareIt-server/src/main/java/ru/practicum/shareit/server/request/ItemRequestRepository.java
@@ -1,7 +1,7 @@
-package ru.practicum.shareit.request;
+package ru.practicum.shareit.server.request;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import ru.practicum.shareit.request.model.ItemRequest;
+import ru.practicum.shareit.server.request.model.ItemRequest;
 
 import java.util.List;
