+++ shareIt-server/src/main/java/ru/practicum/shareit/server/request/dto/ItemRequestShortDto.java
@@ -1,0 +1,11 @@
+package ru.practicum.shareit.server.request.dto;
+
+import lombok.Data;
+
+@Data
+public class ItemRequestShortDto {
+    private Integer id;
+    private String description;
+    private Integer requestorId;
+    private String created;
+}
